+++ ext/Server/Bot/BotAttacking.lua
@@ -243,7 +243,7 @@ local function _DefaultAttackingAction(p_DeltaTime, p_Bot)
 								(p_Bot._ShootModeTimer >= (s_TargetTimeValueRocket - p_DeltaTime - 0.001)) and
 								p_Bot.m_PrimaryGadget ~= nil and p_Bot.m_PrimaryGadget.type == WeaponTypes.Rocket and
 								p_Bot._RocketCooldownTimer <= 0.0 and
-								m_Utilities:CheckProbablity(s_ProbabilityRocket)
+								m_Utilities:CheckProbability(s_ProbabilityRocket)
 							then
 								p_Bot._WeaponToUse = BotWeapons.Gadget1
 							end
@@ -265,7 +265,7 @@ local function _DefaultAttackingAction(p_DeltaTime, p_Bot)
 						(p_Bot.m_Player.soldier.weaponsComponent.weapons[7] and p_Bot.m_Player.soldier.weaponsComponent.weapons[7].primaryAmmo > 0) and
 						p_Bot._ActiveAction ~= BotActionFlags.GrenadeActive) or Config.BotWeapon == BotWeapons.Grenade then
 					-- Should be triggered only once per fireMode.
-					if m_Utilities:CheckProbablity(s_ProbabilityGrenade) then
+					if m_Utilities:CheckProbability(s_ProbabilityGrenade) then
 						if p_Bot.m_Grenade ~= nil
 							and p_Bot._DistanceToPlayer > Registry.BOT.MIN_DISTANCE_NADE
 							and p_Bot._DistanceToPlayer < 25.0 then -- Algorithm only works for up to 25 m.