+++ ext/Server/UIServer.lua
@@ -427,6 +427,13 @@ function FunBotUIServer:_onBotEditorEvent(p_Player, p_Data)
 		end
 		m_NodeCollection:Save(p_Player.name)
 		return
+	elseif request.action == 'refresh_all_waypoints' then
+		if PermissionManager:HasPermission(p_Player, 'UserInterface.WaypointEditor.SaveLoad') == false then
+			ChatManager:SendMessage('You have no permissions for this action.', p_Player)
+			return
+		end
+		m_NodeEditor:RefreshWaypointsOnClient()
+		return
 	elseif request.action == 'waypoints_show_spawns' then
 		if PermissionManager:HasPermission(p_Player, 'UserInterface.WaypointEditor.View') == false then
 			ChatManager:SendMessage('You have no permissions for this action.', p_Player)
@@ -606,7 +613,6 @@ function FunBotUIServer:_writeSettings(p_Player, p_Request)
 	local temporary = false
 	local updateBotTeamAndNumber = false
 	local updateWeaponSets = false
-	local resetSkill = false
 	local calcYawPerFrame = false
 	local updateLanguage = false
 	local updateMaxBots = false
@@ -695,8 +701,6 @@ function FunBotUIServer:_writeSettings(p_Player, p_Request)
 			if s_Changed then
 				if l_Item.UpdateFlag == UpdateFlag.WeaponSets then
 					updateWeaponSets = true
-				elseif l_Item.UpdateFlag == UpdateFlag.Skill then
-					resetSkill = true
 				elseif l_Item.UpdateFlag == UpdateFlag.YawPerSec then
 					calcYawPerFrame = true
 				elseif l_Item.UpdateFlag == UpdateFlag.AmountAndTeam then
@@ -732,10 +736,6 @@ function FunBotUIServer:_writeSettings(p_Player, p_Request)
 	-- Update Weapons if needed.
 	if updateWeaponSets then
 		WeaponList:UpdateWeaponList()
-	end
-
-	if resetSkill then
-		BotManager:ResetSkills()
 	end
 
 	if calcYawPerFrame then