+++ ext/Shared/Utilities.lua
@@ -166,8 +166,19 @@ end
 
 ---@param p_Value number
 ---@return boolean
-function Utilities:CheckProbablity(p_Value)
+function Utilities:CheckProbability(p_Value)
 	return MathUtils:GetRandomInt(1, 100) <= p_Value
+end
+
+---@param p_Angle number
+---@return number
+function Utilities:NormalizeAngleRad(p_Angle)
+	if p_Angle > math.pi then
+		p_Angle = p_Angle - 2 * math.pi
+	elseif p_Angle < -math.pi then
+		p_Angle = p_Angle + 2 * math.pi
+	end
+	return p_Angle
 end
 
 function Utilities:has(p_Object, p_Value)